clase sphreDistance
